+++ src/idtrackerai/tracker/assigner.py
@@ -3,7 +3,7 @@
 import logging
 
 import numpy as np
-from torch import nn
+from torch import load, nn
 
 from idtrackerai import Fragment, ListOfFragments
 from idtrackerai.network import NetworkParams
@@ -73,6 +73,28 @@ def assign_remaining_fragments(
 
     """
     timer.start()
+
+    # Load the penultimate accumulation step if the validation accuracy of the last step was lower
+    if network_params.penultimate_model_path.is_file():
+        current_accuracy = load(network_params.model_path).get("vel_acc", 0.0)
+        penultimate_model = load(network_params.penultimate_model_path)
+        if penultimate_model.get("vel_acc", -1.0) > current_accuracy:
+            logging.info(
+                "The last accumulation step had a lower accuracy than the penultimate."
+            )
+            logging.info(
+                "Loading penultimate model, %s", network_params.penultimate_model_path
+            )
+            identification_model.load_state_dict(penultimate_model, strict=True)
+        else:
+            logging.info(
+                "The last accumulation step had a higher accuracy than the penultimate."
+            )
+    else:
+        logging.warning(
+            "Penultimate model not found (%s)", network_params.penultimate_model_path
+        )
+
     logging.info("Assigning identities to all non-accumulated individual fragments")
     list_of_fragments.reset(roll_back_to="accumulation")
     number_of_unidentified_individual_fragments = (